+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+### Added
+- Aggiunto metodo per verifica fattura con schema xsd dell'sdi
+
 ## [1.1.0]
 ### Added
 - Righe fattura con aliquota diversa